Plein de trucs : branchement des plugin et dirigeables sur les arbres de données, suppression de menus caduques, amélioration du système de traces, suppression des traces en fichiers XML, AnyData sur les arbres de données, corrections de fuites mémoire

+++ sources/src/test/data/DataTreeTest.h
@@ -54,7 +54,7 @@ class DataTreeTest : public Test {
 	const int valeurIntServerId = 2;
 	vector<int> valeurIntServerFullId;
 	const string valeurIntServerName = "valeur-int-server";
-	const float valeurIntServerValue = 637;
+	const int valeurIntServerValue = 637;
 
 	// Server valeur float
 	const int valeurFloatServerId = 3;
